feat(codegen): share error→solution hints with the standalone CLI
Move the Lunora error→solution rule table out of @lunora/vite into a
framework-agnostic module in @lunora/codegen (findLunoraSolution +
LUNORA_SOLUTION_RULES), next to the throw sites that produce the messages.

@lunora/vite's overlay finder now delegates to the shared table instead of
owning its own copy, and the standalone `lunora dev` CLI prints the matched
hint to the terminal under a failing codegen run — so the non-Vite path gets
the same actionable fixes the browser overlay shows, lightly de-marked since
the terminal has no Markdown renderer.
